Added custom callback uri feature
There was a need to be able to set a callback uri that was not using the
same domain as the tenant.

Use-case
The app supports x tenants, via config, and you don't want add x intent
filters to the manifest.
Instead we want to have one "generic" callback uri that we can use
regardless which tenant that was used.